feat:接口增加token验证、新增中行活动获取分享列表接口

See merge request miotech-application/backend/mp2c-go!19

+++ model/request/user.go
@@ -1,5 +1,5 @@
 package request
 
 type ApiUserId struct {
-	ID int `form:"id" binding:"gt=0" alias:"用户id"`
+	ID int64 `form:"id" binding:"gt=0" alias:"用户id"`
 }
